ADDED: SkillManager.SwitchRegister-Enumerator. Does not fully work yet.
Switch Register should move one panel to the side and back and also give it the correct siblingIndex. It does not change the bool used for recognizing which panel should be moved. Some variables like the animation curve are not assigned. This commit is just for being able ot continue at home.